added dummy data for date and time

+++ app/src/main/java/com/example/administrator/earthquakehazard/EarthQuakeAdapterModel.java
@@ -6,26 +6,27 @@ package com.example.administrator.earthquakehazard;
 
 public class EarthQuakeAdapterModel {
 
-        private Double  magnitude;
-        private String location;
-        private int date_time;
+    private Double magnitude;
+    private String location;
+    private Long date_time;
 
-        public EarthQuakeAdapterModel(Double magnitude, String location, int date_time) {
-            this.magnitude = magnitude;
-            this.location = location;
-            this.date_time = date_time;
+    public EarthQuakeAdapterModel(Double magnitude, String location, Long date_time) {
+        this.magnitude = magnitude;
+        this.location = location;
+        this.date_time = date_time;
 
-        }
+    }
 
-        public Double getMagnitude() {
-            return magnitude;
-        }
+    public Double getMagnitude() {
+        return magnitude;
+    }
 
-        public String getLocation() {
-            return location;
-        }
+    public String getLocation() {
+        return location;
+    }
 
-        public int getDate_time() {
-            return date_time;
-        }
+    public Long getDate_time() {
+        return date_time;
+    }
+
 }
